multiple fixes for fleets

+++ app/Services/Game/Formulas/FleetsService.php
@@ -35,7 +35,7 @@ class FleetsService
         return 5;
     }
 
-    public function missionDuration(int $percentage, int $maxFleetSpeed, int $distance, int $speedFactor): float
+    public function missionDuration(int $percentage, float $maxFleetSpeed, int $distance, int $speedFactor): float
     {
         return (35000 / $percentage * sqrt($distance * 10 / $maxFleetSpeed) + 10) / $speedFactor;
     }
